Missing import statement
Added a missing import statement to a solutions file
Also fixed a bug, as the existing version of getConferencesToAttend
used a private version of getUserId() and replaced it with the User API
version. This relates to a workaround for Android which can be ignored
at this point in the course.